Issue with overdue books fixed
There was an issue where the date that the book was checked out wasn't
being cleared out after the book was checked in.  This was causing the
book to show up in the Overdue Books list even when it wasn't even
checked out by anyone.

I also added some test patron and book files that I have been using